crud de convenios hecho
