+++ src/components/agreementUserTable/AgreementUserTable.jsx
@@ -1,15 +1,7 @@
 import styles from "../../pages/agreementAdminPage/styles";
 import PropTypes from "prop-types";
 
-import editIcon from "../../assets/icons/editIcon.svg";
-import deleteIcon from "../../assets/icons/deleteIcon.svg";
-
-function AgreementTable({
-  setOpen,
-  setAgreementId,
-  setOpenDelete,
-  agreements
-}) {
+function AgreementUserTable({ agreements }) {
   return (
     <table className="w-full text-left table-auto border-collapse md:table">
       <thead className="hidden md:table-header-group">
@@ -19,7 +11,6 @@ function AgreementTable({
           <th className={`${styles.thIn} w-[200px]`}>Codigo</th>
           <th className={`${styles.thIn} w-[350px]`}>Descripción</th>
           <th className={`${styles.thIn} w-[200px]`}>Fecha de inicio</th>
-          <th className={`${styles.thOut} w-[250px]`}>Acciones</th>
         </tr>
       </thead>
       <tbody>
@@ -39,7 +30,7 @@ function AgreementTable({
               {agreement.institution}
             </td>
             <td className={styles.tdIn}>
-              <span className="md:hidden font-bold">Codigo: </span>
+              <span className="md:hidden font-bold">Cosido: </span>
               {agreement.agreementNumber}
             </td>
             <td className={styles.tdIn}>
@@ -50,37 +41,6 @@ function AgreementTable({
               <span className="md:hidden font-bold">Fecha de inicio: </span>
               {agreement.startDate}
             </td>
-            <td className={styles.tdOut}>
-              <div className="flex md:justify-center items-center">
-                <span className="md:hidden font-bold">Acciones:</span>
-                <div className="flex md:gap-0 justify-around w-[80%]">
-                  <button
-                    onClick={() => {
-                      setOpen(true);
-                      setAgreementId(agreement.id.toString());
-                    }}
-                  >
-                    <img
-                      className={styles.buttonAction}
-                      src={editIcon}
-                      alt="editIcom"
-                    />
-                  </button>
-                  <button
-                    onClick={() => {
-                      setOpenDelete(true);
-                      setAgreementId(agreement.id.toString());
-                    }}
-                  >
-                    <img
-                      className={styles.buttonAction}
-                      src={deleteIcon}
-                      alt="deleteIcon"
-                    />
-                  </button>
-                </div>
-              </div>
-            </td>
           </tr>
         ))}
       </tbody>
@@ -88,11 +48,8 @@ function AgreementTable({
   );
 }
 
-AgreementTable.propTypes = {
-  agreements: PropTypes.array,
-  setOpen: PropTypes.func,
-  setAgreementId: PropTypes.func,
-  setOpenDelete: PropTypes.func,
-};
+AgreementUserTable.propTypes = {
+    agreements: PropTypes.array.isRequired,
+    };
 
-export default AgreementTable;
+export default AgreementUserTable;
