+++ src/routes/Router.jsx
@@ -4,9 +4,9 @@ import Login from "../pages/loginPage/Login";
 import MainAdmin from "../pages/mainAdminPage/MainAdmin";
 import MainUser from "../pages/mainUserPage/MainUser";
 import FormPage from "../pages/formPage/formPage";
-import ConvenioAdminPage from "../pages/convenioAdminPage/ConvenioAdminPage";
-import CreateConvenioPage from "../pages/createConvenioPage/CreateConvenioPage";
-import ConvenioUserPage from "../pages/convenioUserPage/ConvenioUserPage";
+import AgreementAdminPage from "../pages/agreementAdminPage/AgreementAdminPage";
+import CreateAgreementPage from "../pages/createAgreementPage/CreateAgreementPage";
+import AgreementUserPage from "../pages/agreementUserPage/AgreementUserPage";
 
 function Router() {
     return (
@@ -16,9 +16,9 @@ function Router() {
             <Route path="/admin" element={<MainAdmin />} />
             <Route path="/user" element={<MainUser />} />
             <Route path="/form" element={<FormPage />} />
-            <Route path="/admin/convenio" element={<ConvenioAdminPage />} />
-            <Route path="/admin/convenio/create" element={<CreateConvenioPage />} />
-            <Route path="/user/convenio" element={<ConvenioUserPage />} />
+            <Route path="/admin/convenio" element={<AgreementAdminPage />} />
+            <Route path="/admin/convenio/create" element={<CreateAgreementPage />} />
+            <Route path="/user/convenio" element={<AgreementUserPage />} />
         </Routes>
         </BrowserRouter>
     );
