+++ src/services/agreement.service.js
@@ -1,22 +1,46 @@
 import axios from "./axios.service.js";
 import { apiUrl } from "./env.service.js";
 
+const url = `${apiUrl}/agreement`;
+
 export const getAgreements = async () => {
-    return await axios.get(`${apiUrl}/convenio`);
+    return await axios.get(`${url}/allAgreement`);
 }
 
 export const createAgreement = async (agreement) => {
-    return await axios.post(`${apiUrl}/convenio`, agreement);
+    return await axios.post(`${url}/create`, agreement);
 }
 
 export const updateAgreement = async (agreement, agreementId) => {
-    return await axios.put(`${apiUrl}/convenio/${agreementId}`, agreement);
+    return await axios.put(`${url}/update/${agreementId}`, agreement);
 }
 
 export const deleteAgreement = async (agreementId) => {
-    return await axios.delete(`${apiUrl}/convenio/${agreementId}`);
+    return await axios.delete(`${url}/delete/${agreementId}`);
 }
 
 export const getAgreement = async (agreementId) => {
-    return await axios.get(`${apiUrl}/convenio/${agreementId}`);
+    return await axios.get(`${url}/${agreementId}`);
+}
+
+export const obtainAgreements = async () => {
+    let agreements = await getAgreements();
+
+    agreements = agreements.data;
+
+    const agreementsData = {
+        NATIONAL: [],
+        INTERNATIONAL: [],
+        ALL: agreements
+    }
+
+    agreements.forEach(agreement => {
+        if(agreement.scope === "NATIONAL") {
+            agreementsData.NATIONAL.push(agreement);
+        } else {
+            agreementsData.INTERNATIONAL.push(agreement);
+        }
+    });
+
+    return agreementsData;
 }
